% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ADPIRA.R
\docType{data}
\name{ADPIRA}
\alias{ADPIRA}
\title{ADPIRA}
\format{
Information about the fornat
}
\source{
UMB.
}
\usage{
ADPIRA
}
\description{
Analysis Dataset for the study of Progression Independent of Relapse Activity
in the Polish RRMS patient population.
}
\examples{
attr(ADPIRA, "metadata")
sapply(ADPIRA, attr, which = "label")
head(ADPIRA)
}
\keyword{datasets}
